Fix tutor7pp bug when calling an ausgab object
Fix a tutor7pp bug related to ausgab calls. The value of whether or not
to do russian roulette was not initialized properly. When an ausgab
object was included in the input file, results differed due to random
memory accesses. The results were correct when no ausgab objects were
included.

This is therefore a critical bug fix for tutor7pp.

Also add some ausgab call values that were missing in application.h, for
future use and consistency.